Add README for this directory
This directory doesn't exactly need a README, but my automated build on
Docker Hub is using the README from the last repo it was linked to in
its build details.

It seems like a bug in Docker Hub's process (or rather, this is an edge
case) causes whatever README already existed to stay until another
README is found. The edge case is that I changed the repo linked to the
image from a repo with a README to one without, so the one from the
first repo still shows up in the build details.