Add a puppeteer-friendly image from aws std 4
This is an image based off of the [AWS CodeBuild Standard
4.0](https://github.com/aws/aws-codebuild-docker-images/blob/master/ubuntu/standard/4.0/Dockerfile)
image. The addition to that image in this Dockerfile is that I am adding
the packages that I have found through experimentation to be needed in
order to run an npm package that uses
[Puppeteer](https://developers.google.com/web/tools/puppeteer).

The FROM line in this Dockerfile references a douglasnaphas image, not
an aws one, because AWS does not [make its CodeBuild images
available](https://github.com/aws/aws-codebuild-docker-images/issues/152)
for extending.

The image douglasnaphas/aws-codebuild-docker-images:std4.0 is the [AWS
CodeBuild Standard
4.0](https://github.com/aws/aws-codebuild-docker-images/blob/master/ubuntu/standard/4.0/Dockerfile)
image with no modifications as of
[629ffca](https://github.com/aws/aws-codebuild-docker-images/commit/629ffca6a88d59428354a4bc41eb4ca524dbcbcb).